requirements.txt
clip==1.0
colorama==0.4.6
faiss-cpu==1.11.0
filelock==3.18.0
fsspec==2025.5.1
ftfy==6.3.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
regex==2024.11.6
shapely==2.1.1
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
wcwidth==0.2.13
